Title: 파스칼 삼각형, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 231120/파스칼 삼각형/pascal-triangle.cpp
@@ -3,7 +3,7 @@
 
 using namespace std;
 int main() {
-    int r, c, w;
+    long long r, c, w;
     vector<vector<long long>> pascalTriangle;
 
     long long total = 0;
